requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
fastapi==0.62.0
h11==0.11.0
idna==2.10
iniconfig==1.1.1
mangum==0.10.0
packaging==20.8
pluggy==0.13.1
py==1.10.0
pydantic==1.7.3
pyparsing==2.4.7
pytest==6.2.0
requests==2.25.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.1
Pillow==9.0.0
boto3==1.20.35
python-multipart==0.0.5
aiofiles==0.8.0
tortoise-orm==0.18.1
passlib==1.7.4
bcrypt==3.2.0
PyJWT==2.3.0
gunicorn
geocoder
gunicorn
uvloop
httptools
